+++ FoundationSwaggerTests/Tests/DateTests.swift
@@ -8,6 +8,8 @@
 
 import XCTest
 
+let TIME_TOLERANCE = NSTimeInterval(1e-7)
+
 class DateTests: XCTestCase {
 
     let now = NSDate()
@@ -17,6 +19,11 @@ class DateTests: XCTestCase {
     let positiveInterval = NSTimeInterval(24.42)
     let negativeInterval = NSTimeInterval(-14.99)
     var expectedInterval: NSTimeInterval!
+    var lessThan = false
+    var lessThanOrEqual = false
+    var greaterThan = false
+    var greaterThanOrEqual = false
+
 
     override func setUp() {
         super.setUp()
@@ -24,6 +31,19 @@ class DateTests: XCTestCase {
     
     override func tearDown() {
         super.tearDown()
+    }
+
+    private func makeOtherDates() {
+        earlier = now.dateByAddingTimeInterval(negativeInterval)
+        later = now.dateByAddingTimeInterval(positiveInterval)
+    }
+
+    private func nudgeDateEarlier(date: NSDate) -> NSDate {
+        return date.dateByAddingTimeInterval(-TIME_TOLERANCE)
+    }
+
+    private func nudgeDateLater(date: NSDate) -> NSDate {
+        return date.dateByAddingTimeInterval(TIME_TOLERANCE)
     }
 
     //  MARK: Time interval addition/subtraction
@@ -61,19 +81,70 @@ class DateTests: XCTestCase {
     }
 
     func testSubtractingEarlierDateFromLaterDate() {
-        earlier = now.dateByAddingTimeInterval(negativeInterval)
-        later = now.dateByAddingTimeInterval(positiveInterval)
+        makeOtherDates()
         let interval = later - earlier
         expectedInterval = later.timeIntervalSinceDate(earlier)
         XCTAssertEqual(interval, expectedInterval, "The subtraction operator should produce the time interval between the two dates")
     }
 
     func testSubtractingLaterDateFromEarlierDate() {
-        earlier = now.dateByAddingTimeInterval(negativeInterval)
-        later = now.dateByAddingTimeInterval(positiveInterval)
+        makeOtherDates()
         let interval = earlier - later
         expectedInterval = earlier.timeIntervalSinceDate(later)
         XCTAssertEqual(interval, expectedInterval, "The subtraction operator should produce the time interval between the two dates")
     }
 
+    //  MARK: Date comparison
+
+    func testNowIsEqualToNow() {
+        XCTAssertEqual(now, now, "A date should be considered equal to itself")
+    }
+
+    func testNotNowIsNotEqualToNow() {
+        earlier = nudgeDateEarlier(now)
+        XCTAssertNotEqual(earlier, now, "A date even a smidge earlier should not be considered equal")
+        later = nudgeDateLater(now)
+        XCTAssertNotEqual(later, now, "A date even a smidge later should not be considered equal")
+    }
+
+    func testEarlierDateLessThanNow() {
+        makeOtherDates()
+        lessThan = earlier < now
+        XCTAssertTrue(lessThan, "The earlier date should be considered 'less than' the later date")
+        earlier = nudgeDateEarlier(now)
+        lessThan = earlier < now
+        XCTAssertTrue(lessThan, "The earlier date should be considered 'less than' the later date")
+    }
+
+    func testEarlierDateLessThanOrEqualToNow() {
+        makeOtherDates()
+        lessThanOrEqual = earlier <= now
+        XCTAssertTrue(lessThanOrEqual, "The earlier date should be considered 'less than or equal to' the later date")
+        earlier = nudgeDateEarlier(now)
+        lessThanOrEqual = earlier <= now
+        XCTAssertTrue(lessThanOrEqual, "The earlier date should be considered 'less than or equal to' the later date")
+    }
+
+    func testLaterDateNotLessThanNow() {
+        makeOtherDates()
+        lessThan = later < now
+        XCTAssertFalse(lessThan, "The later date should be considered 'less than' the earlier date")
+    }
+
+    func testLaterDateNotLessThanOrEqualToNow() {
+        makeOtherDates()
+        lessThanOrEqual = later <= now
+        XCTAssertFalse(lessThanOrEqual, "The later date should not be considered 'less than or equal to' the earlier date")
+    }
+    
+    func testNowNotLessThanNow() {
+        lessThan = now < now
+        XCTAssertFalse(lessThan, "A date should not be considered 'less than' itself")
+    }
+
+    func testNowLessThanOrEqualToNow() {
+        lessThanOrEqual = now <= now
+        XCTAssertTrue(lessThanOrEqual, "A date should be considered 'less than or equal to' itself")
+    }
+    
 }