+++ CHANGELOG.md
@@ -5,3 +5,4 @@ FoundationSwagger Changelog
 ##### v0.1.1 (__DATE__)
  - Added NSDate operators for adding/subtracting time intervals
  - Added NSDate operator for producing the interval between two dates
+ - Added NSDate operators for comparing dates